HomePage defined with parishId but style to be improved
The homePage was defined using parishId, change of all fields to english except apv (likely BCM). Style and data for chart to be improved and checked for accuracy with StatisticsService.

Improvement to do on separate CSS per page and cards in HomePage. Clean HomePage with more Pages to ensure readable, light coding and scalability.