I18n integrate and Data Structure modified
The i18n is integrated for 7 languages (in progress). The Data code name are modified to comply with the naming Down -> Up, starting filled from the most upper level to be prefixed by the next down level. Two digits up to 10.000. In a Parish where the Catholic Christians are accounted, the number comes up to 10.000.